+++ common-libs/common/src/services/MetricsService/parsers/index.ts
@@ -1,1 +1,0 @@
-export { VcMetadataParserFactory } from './VcMetadataParser'